package/zl100mt/foris-controller/foris-controller-v0.9.4/requirements.txt
python-prctl
pbkdf2
foris-schema
pytest
pytest-runner
foris-controller-testtools
updater
